Proof of concept of the changes to the config
Idea is that each module handles itself, validates itself using pydantic and returns proper errors. Config only updates when the data is correct, without any errors and only if there are changes
