+++ EyeTrackApp/settings/constants.py
@@ -1,0 +1,1 @@
+BACKGROUND_COLOR = "#424042"
